Use double instead of float everywhere by default
This is mainly important for the drawing code, which achieves higher
precision this way, especially the runtime adjustment routine. To avoid
having to convert between float and double constantly, we also migrate
all the remaining code to double precision.
The only place where we still use floats is when interacting with Java
APIs that support no other data type.

+++ src/main/kotlin/com/loadingbyte/cinecred/common/Picture.kt
@@ -9,49 +9,49 @@ import java.awt.image.BufferedImage
 
 sealed class Picture {
 
-    abstract val width: Float
-    abstract val height: Float
-    abstract fun scaled(scaling: Float): Picture
+    abstract val width: Double
+    abstract val height: Double
+    abstract fun scaled(scaling: Double): Picture
     open fun dispose() {}
 
 
-    class Raster(val img: BufferedImage, val scaling: Float = 1f) : Picture() {
+    class Raster(val img: BufferedImage, val scaling: Double = 1.0) : Picture() {
 
         override val width get() = scaling * img.width
         override val height get() = scaling * img.height
 
-        override fun scaled(scaling: Float) = Raster(img, this.scaling * scaling)
+        override fun scaled(scaling: Double) = Raster(img, this.scaling * scaling)
 
     }
 
 
     class SVG(
-        val gvtRoot: GraphicsNode, private val docWidth: Float, private val docHeight: Float,
-        val scaling: Float = 1f, val isCropped: Boolean = false
+        val gvtRoot: GraphicsNode, private val docWidth: Double, private val docHeight: Double,
+        val scaling: Double = 1.0, val isCropped: Boolean = false
     ) : Picture() {
 
-        override val width get() = scaling * (if (isCropped) gvtRoot.bounds.width.toFloat() else docWidth)
-        override val height get() = scaling * (if (isCropped) gvtRoot.bounds.height.toFloat() else docHeight)
+        override val width get() = scaling * (if (isCropped) gvtRoot.bounds.width else docWidth)
+        override val height get() = scaling * (if (isCropped) gvtRoot.bounds.height else docHeight)
 
-        override fun scaled(scaling: Float) = SVG(gvtRoot, docWidth, docHeight, this.scaling * scaling, isCropped)
+        override fun scaled(scaling: Double) = SVG(gvtRoot, docWidth, docHeight, this.scaling * scaling, isCropped)
         fun cropped() = SVG(gvtRoot, docWidth, docHeight, scaling, isCropped = true)
 
     }
 
 
     class PDF private constructor(
-        private val aDoc: AugmentedDoc, val scaling: Float, val isCropped: Boolean
+        private val aDoc: AugmentedDoc, val scaling: Double, val isCropped: Boolean
     ) : Picture() {
 
-        constructor(doc: PDDocument) : this(AugmentedDoc(doc), 1f, false)
+        constructor(doc: PDDocument) : this(AugmentedDoc(doc), 1.0, false)
 
         val doc get() = aDoc.doc
         val minBox get() = aDoc.minBox
 
-        override val width get() = scaling * (if (isCropped) minBox.width.toFloat() else doc.pages[0].cropBox.width)
-        override val height get() = scaling * (if (isCropped) minBox.height.toFloat() else doc.pages[0].cropBox.height)
+        override val width get() = scaling * (if (isCropped) minBox.width else doc.pages[0].cropBox.width.toDouble())
+        override val height get() = scaling * (if (isCropped) minBox.height else doc.pages[0].cropBox.height.toDouble())
 
-        override fun scaled(scaling: Float) = PDF(aDoc, this.scaling * scaling, isCropped)
+        override fun scaled(scaling: Double) = PDF(aDoc, this.scaling * scaling, isCropped)
         fun cropped() = PDF(aDoc, scaling, isCropped = true)
 
         override fun dispose() {
