+++ src/main/kotlin/com/loadingbyte/cinecred/project/Drawn.kt
@@ -10,6 +10,6 @@ class DrawnPage(val defImage: DeferredImage, val stageInfo: PersistentList<Drawn
 
 sealed class DrawnStageInfo {
     class Card(val middleY: Y) : DrawnStageInfo()
-    class Scroll(val scrollStartY: Y, val ownedScrollHeight: Y, val frames: Int, val initialAdvance: Float) :
+    class Scroll(val scrollStartY: Y, val ownedScrollHeight: Y, val frames: Int, val initialAdvance: Double) :
         DrawnStageInfo()
 }
